fix(storybook): move configuration to TypeScript
